requirements.txt
flask
flask_cors
llama-cpp-python
huggingface-hub
